Move toast to functions file
